Store metrics filters in URL
So that filters are preserved upon page reload or direct linking
Closes https://github.com/kiali/kiali/issues/1558
